requirements.txt
img==2.5
Pillow==9.3.0
PyAutoGUI==0.9.53
streamlit==1.15.2
streamlit_option_menu==0.3.2
